feat: add dist
