+++ package.json
@@ -2,11 +2,10 @@
   "name": "image-list-actions",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "ncc build src/action.js -o dist",
-    "start": "node dist/index.js"
+    "dist": "npx -p @zeit/ncc ncc build index.js",
+    "format": "prettier --write '*.js' --tab-width=4"
   },
   "repository": {
     "type": "git",
